Moving parse datetime to separate test (#97)
Trouble is that parsing DateTime depends on machine configuration and how the tests are executed.
When using F#+ and executing the tests in [a repository](https://github.com/wallymathieu/DateTimeBehaviorOnTravisAndAppveyor) based on the MiniScaffold template we notice that the behavior is consistent. While executing using NUnit2 test runner on mono Mac OS X, mono on Linux image on Travis and .net on Windows image on Appveyor gives different behaviors.
In order to not to fix something in F#+ for which there is a library that [solves DateTime issues](https://github.com/nodatime/nodatime/), better to ignore the discrepancies.